Replaced vagrant username and key by basebox
Using vagrant username gave the wrong impression that we are using
vagrant.  The username and key is configured in the project and
in the platform project and they are therefore easy to change.

If user want to use a prebuilt vagrant image instead, they can
still do it in their project.